move google actions into util
